Refactor skill and experience lists into server and client components
